extract widget
